renamed app-bundle to chunked-bundle
