+++ test/spec/index.js
@@ -6,10 +6,10 @@ import path from "path";
 import combineSourceMap from "combine-source-map";
 import BitBundler from "../../src/index";
 import wrapModule from "../helpers/wrapModule";
-import appBundlePrelude from "../../src/bundle/app-bundle-prelude";
+import chunkedBundlePrelude from "../../src/bundle/chunked-bundle-prelude";
 import fs from "fs";
 
-const prelude = appBundlePrelude.toString();
+const prelude = chunkedBundlePrelude.toString();
 
 describe("BitBundler test suite", function() {
   var createBundler, bitbundler;
